add assets for new game demo, add Dorothy new logo, add VGNode functions.

+++ Source/3rdParty/Box2D/Dynamics/b2Fixture.h
@@ -39,16 +39,16 @@ struct b2Filter
 	}
 
 	/// The collision category bits. Normally you would just set one bit.
-	uint16 categoryBits;
+	uint32 categoryBits;
 
 	/// The collision mask bits. This states the categories that this
 	/// shape would accept for collision.
-	uint16 maskBits;
+	uint32 maskBits;
 
 	/// Collision groups allow a certain group of objects to never collide (negative)
 	/// or always collide (positive). Zero means no collision group. Non-zero group
 	/// filtering always wins against the mask bits.
-	int16 groupIndex;
+	int8 groupIndex;
 };
 
 /// A fixture definition is used to create a fixture. This class defines an
